+++ Proyecto Unity 2D/Assets/Prefabs/GenerdorProyectil.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 543691703154dc343b6c9750950c10f9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
