+++ Proyecto Unity 2D/Assets/Prefabs/Proyectiles/Proyectil Parabolico.prefab
@@ -1,0 +1,91 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &7123644389784621308
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 853347409556970922, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 2.15
+      objectReference: {fileID: 0}
+    - target: {fileID: 853347409556970922, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -3.03
+      objectReference: {fileID: 0}
+    - target: {fileID: 853347409556970922, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 853347409556970922, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 853347409556970922, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 853347409556970922, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 853347409556970922, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 853347409556970922, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 853347409556970922, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 853347409556970922, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4483737553082260240, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_Color.b
+      value: 0.25119257
+      objectReference: {fileID: 0}
+    - target: {fileID: 4483737553082260240, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_Color.g
+      value: 0.25119257
+      objectReference: {fileID: 0}
+    - target: {fileID: 4483737553082260240, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_Color.r
+      value: 0.9056604
+      objectReference: {fileID: 0}
+    - target: {fileID: 8425493659269479894, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      propertyPath: m_Name
+      value: Proyectil Parabolico
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 8425493659269479894, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1311144563646498698}
+  m_SourcePrefab: {fileID: 100100000, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+--- !u!1 &1599086846141345066 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 8425493659269479894, guid: 918b5c53d2f160f4c8facf4d58585fa1, type: 3}
+  m_PrefabInstance: {fileID: 7123644389784621308}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &1311144563646498698
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1599086846141345066}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 804a52cf67e01404981975efb01c245c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
