+++ Proyecto Unity 2D/Assets/Prefabs/Proyectiles/Proyectil Parabolico.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8184bf2bb9f438842b68d3c3fce9adbb
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
