+++ Proyecto Unity 2D/Assets/Prefabs/Proyectiles/Proyectil Recto.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 97436adb50a8ebc4d9b78920b2f9149c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
