+++ Proyecto Unity 2D/Assets/Scripts/Proyectiles/Proyectil.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d439f140110e85740b945cb076223a3d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
