+++ Proyecto Unity 2D/Assets/Scripts/Proyectiles/ProyectilParabolico.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 804a52cf67e01404981975efb01c245c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
