+++ Proyecto Unity 2D/Assets/Scripts/Proyectiles/ProyectilRecto.cs
@@ -2,30 +2,14 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class ProyectilRecto : MonoBehaviour
+public class ProyectilRecto : Proyectil
 {
-    [SerializeField]
-    [Range(1f, 30f)]
-    private float speed = 10f;
 
-    private Rigidbody2D rb;
-    private void Awake()
-    {
-        rb = GetComponent<Rigidbody2D>();
-    }
-
-    private void OnEnable()
-    {
-        Mover();
-    }
-
-    private void Mover()
+    protected override void Mover()
     {
         // Establece la dirección del proyectil (por ejemplo, hacia la izquierda en X)
         Vector2 direction = Vector2.left;
         // Aplica la velocidad al Rigidbody
         rb.velocity = direction * speed;
     }
-
-
 }
